pyproject.toml
[project]
name = "friday-night-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi-pagination>=0.15.10",
    "fastapi[standard]>=0.129.0",
    "gunicorn>=25.1.0",
    "httpx>=0.28.1",
    "openai>=2.24.0",
    "orjson>=3.11.7",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "sqlmodel>=0.0.33",
    "structlog>=25.5.0",
    "supabase>=2.28.0",
    "supabase-auth>=2.28.0",
    "uuid6>=2025.0.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "polyfactory>=3.2.0",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.1",
    "taskipy>=1.14.1",
    "testcontainers>=4.14.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"
